+++ lib/features/home/presentation/home_screen.dart
@@ -4,6 +4,7 @@ import 'package:test1/cubit/loader_cubit/loader_cubit.dart';
 import 'package:test1/cubit/loader_cubit/loader_cubit.dart';
 import 'package:test1/cubit/loader_cubit/loader_state.dart';
 import 'package:test1/features/home/presentation/home_view_model.dart';
+import '../../../components/widgets/loader.dart';
 import '../../../core/constants/constants.dart';
 import '../../../core/constants/vertical_and_horizontal_space.dart';
 import '../../../core/theming/styles.dart';
@@ -114,32 +115,4 @@ class _HomeScreenState extends State<HomeScreen> {
   }
 }
 
-class Loader extends StatelessWidget {
-  const Loader({
-    super.key,
-    required this.loading,
-  });
 
-  final Loading loading;
-
-  @override
-  Widget build(BuildContext context) {
-    return BlocBuilder<Loading, LoadingState>(
-        bloc: loading,
-        builder: (context, state) {
-          if (state.loading) {
-            return  Container(
-                height: 1.sh,
-             width: 1.sw,
-             decoration: BoxDecoration(
-               borderRadius: BorderRadius.circular(7)
-                   ,   color: Colors.black.withOpacity(.1),
-             ),
-                child: const Center(child: CircularProgressIndicator()));
-          }
-          else{
-            return const SizedBox.shrink();
-          }
-        });
-  }
-}
